Refactor clock implementations
