+++ pyrate_limiter/clocks.py
@@ -1,4 +1,8 @@
+"""Clock implementation using different backend
+"""
 import sqlite3
+from pathlib import Path
+from tempfile import gettempdir
 from time import monotonic
 from time import time
 
@@ -18,13 +22,30 @@ class TimeClock(Clock):
         return int(1000 * time())
 
 
+class TimeAsyncClock(Clock):
+    """Time Async Clock, meant for testing only"""
+
+    async def now(self) -> int:
+        return int(1000 * time())
+
+
 class SQLiteClock(Clock):
-    """Get timestamp using SQLite"""
+    """Get timestamp using SQLite as remote clock backend"""
 
     conn: sqlite3.Connection
     time_query = "SELECT CAST(ROUND((julianday('now') - 2440587.5)*86400000) As INTEGER)"
 
-    def __init__(self, conn: sqlite3.Connection):
+    def __init__(self, conn: sqlite3.Connection = None):
+        if conn is None:
+            temp_dir = Path(gettempdir())
+            default_db_path = temp_dir / "pyrate_limiter_clock_only.sqlite"
+
+            conn = sqlite3.connect(
+                default_db_path,
+                isolation_level="EXCLUSIVE",
+                check_same_thread=False,
+            )
+
         self.conn = conn
 
     def now(self) -> int:
