+++ tests/conftest.py
@@ -1,16 +1,15 @@
-import sqlite3
+"""Pytest config
+"""
 from logging import basicConfig
 from logging import getLogger
 from os import getenv
-from pathlib import Path
-from tempfile import gettempdir
-from time import time
+from typing import Union
 
 import pytest
 
-from pyrate_limiter.abstracts import Clock
 from pyrate_limiter.clocks import MonotonicClock
 from pyrate_limiter.clocks import SQLiteClock
+from pyrate_limiter.clocks import TimeAsyncClock
 from pyrate_limiter.clocks import TimeClock
 
 # Make log messages visible on test failure (or with pytest -s)
@@ -19,24 +18,14 @@ basicConfig(level="INFO")
 getLogger("pyrate_limiter").setLevel(getenv("LOG_LEVEL", "INFO"))
 
 
-temp_dir = Path(gettempdir())
-default_db_path = temp_dir / "pyrate_limiter_clock_only.sqlite"
+clocks = [MonotonicClock(), TimeClock(), SQLiteClock(), TimeAsyncClock()]
 
-conn = sqlite3.connect(
-    default_db_path,
-    isolation_level="EXCLUSIVE",
-    check_same_thread=False,
-)
-
-
-class MockAsyncClock(Clock):
-    """Mock Async Clock, only for testing"""
-
-    async def now(self) -> int:
-        return int(1000 * time())
-
-
-clocks = [MonotonicClock(), TimeClock(), SQLiteClock(conn), MockAsyncClock()]
+ClockSet = Union[
+    MonotonicClock,
+    TimeClock,
+    SQLiteClock,
+    TimeAsyncClock,
+]
 
 
 @pytest.fixture(params=clocks)
